Basically finished with all NonFlatBgd stuff. What's new?  Previously, to get the NonFlatBgd, I was simply using TH1::Fit method with a Chi2 fit on the Cf.  Technically, this is not right, although, in practice, the results do not differ (at least when FitType is linear, not sure otherwise).  Now, by default (I still have option to use old Chi2 method) I use a full PML method, with numerators and denominators separately.
Now, for FitPairAnalysis:  Previously, the method was the same as above (obviously CfHeavy instead of CfLite).  Now, I can use either kChi2PML or kChi2.  Also, I have two methods: GetNonFlatBackground_FitCombinedPartials and GetNonFlatBackground_CombinePartialFits (most correct method).  In _FitCombinedPartials, the histograms from the partial analyses are first combined, and the resultant histograms fit.  In the case of kChi2, I simply use fKStarCfHeavy.  In case of kChi2PML, where I need numerators and denominators, I simply add the numerators, and simply add the denominators.  In the case of _CombinePartialFits, the histograms from the partial analyses are first fit, and then combined in a weighted mean (with the numerator counts as the weights).

For FitPairAnalysis methods, this will likely not be used for anything other than plotting the final results (in which the pair analyses are plotted, not partials, so the backgrounds must be combined in some manner).  However, I was previously attempting, in LednickyFitter, to grab the NonFlatBgd for the partial analyses from the pair analyses (in case of Quadratic and Gaussian), in an effort to stabilize the fits.  This actually doesn't seem to make much difference, so I will likely remove this feature.

Finally, when using kChi2 method (i.e. simply TH1::Fit method), when using option "0qWL" (for loglikelihood with non-integer values) instead of "0q", I get different results.  In all permuations of above options (kChi2 vs kChi2PML, using partial vs pair to fit, etc), this difference in "0qWL" vs "0q" is the only thing that causes a difference in the final fit results.  Using the "WL" option with a CF is probably not correct, but still interesting that differences are so significant (even with kLinear form of fit function)!